agents/postmortem_agent/requirements.txt
redis>=4.5.1
openai>=0.27.0
requests>=2.28.0
python-dotenv>=1.0.0
crewai>=0.28.0
langchain-openai>=0.0.2
# Knowledge base dependencies
qdrant-client>=1.6.0
weaviate-client>=3.25.0
markdown>=3.5.0 